comleted 4 points

+++ project/include/utils.h
@@ -2,8 +2,7 @@
 #define PROJECT_INCLUDE_UTILS_H_
 
 int timer_from(unsigned int from);
-int custom_pow(int base, int pow);
+long long int custom_pow(int base, int pow);
   // TODO(Ivan Biryukov): Stuff. Implement `power of` function
-  // int custom_pow(int base, int power);
 
 #endif  // PROJECT_INCLUDE_UTILS_H_
